Readme.md
This is a collection of my revison materials for the Microsfot C# exam. The examples are mainly focused on concepts I have little experience using so the examples are not fully complete in comparison the book provided by microsoft. 
The exercises are a comparison of my answers compared to the C# microsoft book.